+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Release History
 
+### 1.8.0 (Unreleased)
+#### New Features
+* Updated `azure-cosmos` version to 4.42.0.
+* Updated `json-path` version to 2.8.0.
+* Updated `kafka-connect` version to 3.4.0.
+
+#### Key Bug Fixes
+* Fixed issue in Sink Connector for nested maps in Struct types. [PR 508](https://github.com/microsoft/kafka-connect-cosmosdb/pull/508)
+
 ### 1.7.1 (2023-03-24)
 #### Key Bug Fixes
 * Fixed issue in Sink Connector when using `AvroConverter`. [PR 503](https://github.com/microsoft/kafka-connect-cosmosdb/pull/503)